+++ CHANGELOG.md
@@ -8,6 +8,7 @@ types you need to work with from the rest of your code.
 ### Changed
 
 - Enable strict type checking in analyzer
+- Renamed Celcius to Celsius (correct spelling) per https://github.com/mongoose13/fling-units/issues/2
 
 ## [3.0.5]